Updated to handle a few edge cases, especially very long titles, from mangling the og:image. Also refactored the docs route function so that it is not duplicated functionality between 2 plugins. Moved the background so that it is not deployed to the public site.
